raspberry/requirements.txt
pynmea2
RPi.GPIO